Convert validation errors to core.ValidationError
This commit changes the functionality of the model validation methods to
instead return the new core.ValidationError type rather than raw error
strings. In order to accomodate the new types for testing, the test
runner for validations was updated.

I also made the decision to split the tests for valid models into
separate functions since not doing so complicated the logic for the test
runner unnecessarily.

Incidentally, I also had to change the Model interface since the Name
method quickly clashed with Organization.Name.