Add authorization error type
This commit also updates the authz service methods to return
authorization errors where appropriate. For method calls that are not
directly related to authorization, the service returns the bare error it
receives. This is intentional since failing to retrieve a record isn't
really indiciative of an authorization failure. It's not ideal though
since at minimum we would want to wrap those errors. They can be updated
later when I add some sort of error wrap mechanism.