Add raw error type and return a raw error for expired invitations
The raw error type is primarily used for situations where you need to
return a bespoke error message to be useful, but there isn't a relevant
error type already and it would be overkill to add a new type. In this
commit, we're returning a raw error when an invitation is expired at the
time someone tries to accept it. Creating a new type here would be
ridiculous since it wouldn't be used anywhere else, but the code and
message are specific to this use-case so a "bad request" error wouldn't
fit either.