Replace in-code sql with embedded sql
By embedding sql into strings, I get the benefit of autocomplete and
syntax highlighting in SQL tooling as well as the ability to run the
queries directly without having to reformat them.

This commit also changes the terminology at the repository layer to
match the database naming so Create becomes Insert and Save becomes
Update. This is just to ensure that the SQL query, its name, and the
method calling it all use the same terminology.